requirements.txt
# Flask Framework
Flask==2.2.3
click==8.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.3

# Flask Packages
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Flask-User==1.0.1.5
flask_bcrypt==1.0.1

# MySQL
mysql-connector-python==8.0.28
paramiko==3.5.1
wikipedia==1.4.0
